requirements.txt
django-extensions==1.5.2
